+++ internal/aghtime/duration_test.go
@@ -1,4 +1,4 @@
-package home
+package aghtime
 
 import (
 	"bytes"